@@ -7,8 +7,8 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
-        classpath 'com.google.gms:google-services:3.1.0'
+        classpath 'com.android.tools.build:gradle:3.2.0'
+        classpath 'com.google.gms:google-services:3.2.0'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -30,6 +30,6 @@ task clean(type: Delete) {
 }
 
 ext {
-    compileSdkVersion = 27
-    supportLibVersion = "27.0.2"
+    compileSdkVersion = 28
+    supportLibVersion = "28.0.0"
 }
