@@ -1,8 +1,8 @@
 package com.community.jboss.leadmanagement;
 
-import android.arch.persistence.room.Room;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.room.Room;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import com.community.jboss.leadmanagement.data.LeadDatabase;
 import com.community.jboss.leadmanagement.data.entities.Contact;
