@@ -2,7 +2,7 @@ package com.community.jboss.leadmanagement;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.community.jboss.leadmanagement.main.MainActivity;
 
