@@ -1,6 +1,6 @@
 package com.community.jboss.leadmanagement.main.contacts.importcontact;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
