@@ -1,7 +1,7 @@
 package com.community.jboss.leadmanagement.main.groups;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
