@@ -1,8 +1,8 @@
 package com.community.jboss.leadmanagement.main;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 
 
 public class MainActivityViewModel extends ViewModel {
